chore: Move layer folder to up-level directory
