+++ packages/artalk/src/artalk.ts
@@ -9,9 +9,9 @@ import defaults from './defaults'
 import CheckerLauncher from './lib/checker'
 import Editor from './editor'
 import List from './list'
-import SidebarLayer from './components/sidebar-layer'
+import SidebarLayer from './layer/sidebar-layer'
 
-import Layer, { GetLayerWrap } from './components/layer'
+import Layer, { GetLayerWrap } from './layer'
 import Api from './api'
 import * as Utils from './lib/utils'
 import * as Ui from './lib/ui'
