+++ packages/artalk/src/layer/index.ts
@@ -1,0 +1,4 @@
+import Layer, { GetLayerWrap } from './layer'
+
+export default Layer
+export { GetLayerWrap }
