+++ packages/artalk/src/lib/checker/index.ts
@@ -1,7 +1,7 @@
 import Context from '~/types/context'
-import Layer from '@/components/layer'
-import Dialog from '@/components/dialog'
 import { CheckerPayload } from '~/types/event'
+import Dialog from '@/components/dialog'
+import Layer from '@/layer'
 import * as Utils from '../utils'
 import * as Ui from '../ui'
 import CaptchaChecker from './captcha-checker'
